+++ packages/runtime/src/widgets/Form/FormItem.tsx
@@ -18,11 +18,12 @@ export const EnsembleFormItem: React.FC<EnsembleFormItemProps<unknown>> = (
   const { values, rules, ...rest } = props;
   const { backgroundColor: _, ...formItemStyles } = values?.styles ?? {};
   const formInstance = AntForm.useFormInstance();
+  const requiredRule = { required: Boolean(values?.required) };
 
   return (
     <AntForm.Item
       className={values?.styles?.names}
-      initialValue={values?.value}
+      initialValue={values?.initialValue}
       label={
         values?.label ? (
           <label
@@ -39,7 +40,7 @@ export const EnsembleFormItem: React.FC<EnsembleFormItemProps<unknown>> = (
         ) : null
       }
       name={formInstance ? values?.id ?? values?.label : undefined}
-      rules={rules?.concat([{ required: Boolean(values?.required) }])}
+      rules={rules?.concat(requiredRule) || [requiredRule]}
       validateTrigger={
         values?.validateOnUserInteraction === true ? "onChange" : "onSubmit"
       }